third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/mask-image-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'mask-image' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'mask-image' to an image
  assert_class_string: Image value must be a CSSImageValue expected "[object CSSImageValue]" but got "[object CSSStyleValue]"
Harness: the test ran to completion.
